Cleaned up correlations and memory kernel implementation
The correlation function has been split into its imaginary and real
part. This is useful since they usually appear independently in open
system dynamics equations. Furthermore, for many spectral densities the
imaginary part can be computed analytically while the real is hard to
integrate even in some of the simplest cases.

Finally, the memory kernel has been re-expressed in terms of the
imaginary part of the correlation function to save on code duplication.